backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pymongo>=4.5.0
python-multipart>=0.0.9
pydantic>=2.6.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
